+++ src/ovl5.c
@@ -57,8 +57,8 @@ void func_ovl5_8018D0C0()
 {
 	lbRelocSetup rldm_setup;
 
-	rldm_setup.table_addr = (s32)&lLBRelocTableAddr;
-	rldm_setup.table_files_num = (u32)&lLBRelocTableFilesNum;
+	rldm_setup.table_addr = &lLBRelocTableAddr;
+	rldm_setup.table_files_num = &lLBRelocTableFilesNum;
 	rldm_setup.file_heap = NULL;
 	rldm_setup.file_heap_size = 0;
 	rldm_setup.status_buf = D_ovl5_8018D678;
